# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple billing system",
        "title": "Billing System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/api/operations/": {
            "get": {
                "description": "Get wallet operations logs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "operations"
                ],
                "summary": "Wallet operations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Report format",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "date",
                        "in": "query"
                    }
                ]
            }
        },
        "/api/users/": {
            "post": {
                "description": "Create new user and wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create new user",
                "parameters": [
                    {
                        "description": "User attributes",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.UserForm"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create user response",
                        "schema": {
                            "$ref": "#/definitions/serializers.UserSerializer"
                        }
                    },
                    "400": {
                        "description": "User form validation error",
                        "schema": {
                            "$ref": "#/definitions/http.FormErrorSerializer"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorMsg"
                        }
                    }
                }
            }
        },
        "/api/users/{id}/enroll/": {
            "post": {
                "description": "Enroll particular users wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Enroll wallet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Enrollment attributes",
                        "name": "enroll",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.EnrollForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieving user information with updated balance",
                        "schema": {
                            "$ref": "#/definitions/serializers.UserSerializer"
                        }
                    },
                    "400": {
                        "description": "Enroll form validation error",
                        "schema": {
                            "$ref": "#/definitions/http.FormErrorSerializer"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorMsg"
                        }
                    }
                }
            }
        },
        "/api/wallets/transfer/": {
            "post": {
                "description": "Transfer funds between two users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Transfer funds",
                "parameters": [
                    {
                        "description": "Transfer parameters",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.WalletForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wallet from id",
                        "schema": {
                            "$ref": "#/definitions/serializers.WalletSerializer"
                        }
                    },
                    "400": {
                        "description": "Wallet transfer validation error",
                        "schema": {
                            "$ref": "#/definitions/http.FormErrorSerializer"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorMsg"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "forms.EnrollForm": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                }
            }
        },
        "forms.UserForm": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "forms.WalletForm": {
            "type": "object",
            "required": [
                "amount",
                "wallet_from",
                "wallet_to"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "wallet_from": {
                    "type": "integer"
                },
                "wallet_to": {
                    "type": "integer"
                }
            }
        },
        "http.ErrorMsg": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "http.FormErrorSerializer": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "serializers.UserSerializer": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "serializers.WalletSerializer": {
            "type": "object",
            "properties": {
                "wallet_from": {
                    "type": "integer"
                }
            }
        }
    }
}